# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/transactions": {
      "post": {
        "operationId": "TransactionsController_createTransaction",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      },
      "get": {
        "operationId": "TransactionsController_findAllTransactions",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/{id}": {
      "get": {
        "operationId": "TransactionsController_findTransaction",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      },
      "patch": {
        "operationId": "TransactionsController_updateTransaction",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      },
      "delete": {
        "operationId": "TransactionsController_deleteTransaction",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/{id}/sync": {
      "post": {
        "operationId": "TransactionsController_syncTransaction",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/{id}/items": {
      "get": {
        "operationId": "TransactionsController_getTransactionItems",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      }
    },
...
  },
  "info": {
    "title": "SabiMarket API",
    "description": "Backend API for the SabiMarket AI voice-first market assistant.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateTransactionItemDto": {
        "type": "object",
        "properties": {
          "itemName": {
            "type": "string",
            "example": "Orange"
          },
          "quantity": {
            "type": "number",
            "example": 5
          },
          "unit": {
            "type": "string",
            "example": "piece"
          },
          "unitPrice": {
            "type": "number",
            "example": 500
          }
        },
        "required": [
          "itemName",
          "quantity",
          "unit",
          "unitPrice"
        ]
      },
      "CreateTransactionDto": {
        "type": "object",
        "properties": {
          "transactionType": {
            "type": "string",
            "example": "buy",
            "description": "e.g. buy, sell, debt_owed, debt_paid, expense, waste"
          },
          "totalAmount": {
            "type": "number",
            "example": 42500
          },
          "profit": {
            "type": "number",
            "example": 0
          },
          "currency": {
            "type": "string",
            "example": "NGN"
          },
          "note": {
            "type": "string",
            "example": "Bought stock from Mile 12"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateTransactionItemDto"
            }
          }
        },
        "required": [
          "transactionType",
          "totalAmount",
          "items"
        ]
      },
      "UpdateTransactionDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "description": "Only the note can be edited after creation. Editing type/amount/items is not supported because it would desynchronize inventory."
          }
        }
      },
...
    }
  }
}